@@ -1,11 +1,14 @@
 {
+  "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "rimraf ./build && tsc",
+    "dev": "npm run build && npm run swagger && npm run start",
     "start": "node build/index.js",
+    "build": "rimraf ./build && tsc",
     "docker_build": "docker build -t badbeatmods .",
     "build_start": "npm run build && npm run start",
-    "swagger": "ts-node src/api/generateSwagger.ts"
+    "swagger": "npx tsx src/api/generateSwagger.ts",
+    "test": "vitest run",
+    "lint": "npx eslint ./src --ext .ts"
   },
   "dependencies": {
     "@octokit/rest": "^21.1.0",
@@ -36,13 +39,14 @@
     "zod": "^3.24.1"
   },
   "devDependencies": {
+    "@faker-js/faker": "^9.5.1",
     "@octokit/types": "^13.7.0",
     "@types/connect-sqlite3": "^0.9.5",
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
     "@types/express-fileupload": "^1.5.1",
     "@types/express-session": "^1.18.0",
-    "@types/node": "^20.16.9",
+    "@types/node": "^20.17.23",
     "@types/node-fetch": "^2.6.11",
     "@types/passport": "^1.0.17",
     "@types/passport-discord": "^0.1.14",
@@ -50,13 +54,16 @@
     "@types/passport-http-bearer": "^1.0.41",
     "@types/pg": "^8.11.10",
     "@types/semver": "^7.5.8",
+    "@types/supertest": "^6.0.2",
     "@types/swagger-ui-express": "^4.1.7",
     "@typescript-eslint/eslint-plugin": "^7.18.0",
     "@typescript-eslint/parser": "^7.18.0",
     "eslint": "^8.57.1",
     "rimraf": "^5.0.10",
+    "supertest": "^7.0.0",
     "swagger-autogen": "^2.23.7",
     "ts-node": "^10.9.2",
-    "typescript": "^5.6.2"
+    "typescript": "^5.8.2",
+    "vitest": "^3.0.7"
   }
 }
