@@ -1,7 +1,7 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-import { Categories, Mod, Platform, Status, UserRoles } from '../shared/Database';
+import { Categories, Platform, Status, UserRoles } from '../shared/Database.ts';
 import swaggerAutogen from 'swagger-autogen';
-import { OpenAPIV3, OpenAPIV3_1 } from 'openapi-types';
+import { OpenAPIV3_1 } from 'openapi-types';
 
 // docs: https://swagger-autogen.github.io/docs/getting-started/quick-start/
 const options = {
@@ -36,7 +36,7 @@ const DBObject: OpenAPIV3_1.SchemaObject = {
             example: `2023-10-01T00:00:00.000Z`,
         },
     }
-}
+};
 
 const ModDBObject: OpenAPIV3_1.SchemaObject = {
     type: `object`,
@@ -475,7 +475,7 @@ const doc = {
     }
 };
 
-const outputFile = `./swagger.json`;
+const outputFile = `./swagger_full.json`;
 const routes = [
     `./routes/beatmods.ts`,
     `./routes/getMod.ts`,
@@ -488,7 +488,7 @@ const routes = [
     `./routes/approval.ts`,
     `./routes/motd.ts`,
     `./routes/users.ts`,
-    `./routes/status.ts`,
+    `./routes/apistatus.ts`,
     `./routes/bulkActions.ts`,
 ];
 
