@@ -1,8 +1,8 @@
 import { Router } from 'express';
-import swaggerDocument from '../../api/swagger.json';
-import * as fs from 'fs';
-import { validateSession } from '../../shared/AuthHelper';
-import { User } from 'src/shared/Database';
+import { validateSession } from '../../shared/AuthHelper.ts';
+import { User } from '../../shared/Database.ts';
+import { Config } from '../../shared/Config.ts';
+import { Utils } from '../../shared/Utils.ts';
 
 export class StatusRoutes {
     private router: Router;
@@ -42,25 +42,8 @@ export class StatusRoutes {
     }
 
     private generateStatusResponse(user: User | null) {
-        let gitVersion = `Version not found.`;
-        let apiVersion = `Version not found.`;
-        if (fs.existsSync(`.git/HEAD`) || process.env.GIT_VERSION) {
-            if (process.env.GIT_VERSION) {
-                gitVersion = `${process.env.GIT_VERSION.substring(0, 7)}`;
-            } else {
-                let gitId = fs.readFileSync(`.git/HEAD`, `utf8`);
-                if (gitId.indexOf(`:`) !== -1) {
-                    let refPath = `.git/` + gitId.substring(5).trim();
-                    gitId = fs.readFileSync(refPath, `utf8`);
-                }
-
-                gitVersion = `${gitId.substring(0, 7)}`;
-            }
-        }
-
-        if (swaggerDocument?.info?.version) {
-            apiVersion = `${swaggerDocument.info.version}`;
-        }
+        let gitVersion = Utils.getGitVersion();
+        let apiVersion = Config.API_VERSION;
 
         let message = `API is running.`;
         if (user) {
