@@ -1,10 +1,9 @@
 import { Router } from 'express';
-import { DatabaseHelper, EditQueue, ModVersion, Status, UserRoles } from '../../shared/Database';
-import { validateSession } from '../../shared/AuthHelper';
-import { Validator } from '../../shared/Validator';
+import { DatabaseHelper, EditQueue, ModVersion, Status, UserRoles } from '../../shared/Database.ts';
+import { validateSession } from '../../shared/AuthHelper.ts';
+import { Validator } from '../../shared/Validator.ts';
 import { Op } from 'sequelize';
-import { Logger } from '../../shared/Logger';
-import { SemVer } from 'semver';
+import { Logger } from '../../shared/Logger.ts';
 
 export class BulkActionsRoutes {
     private router: Router;
@@ -91,7 +90,7 @@ export class BulkActionsRoutes {
             };
 
             for (let modVersion of modVersions) {
-                let outObj = await modVersion.addGameVersionId(gameVersion.id, session.user.id).catch((err) => {
+                let outObj = await modVersion.addGameVersionId(gameVersion.id, session.user).catch((err) => {
                     Logger.error(`Error adding game version ${gameVersion.id} to mod version ${modVersion.id}: ${err}`);
                     //results.errorIds.push(modVersion.id);
                     return null;
@@ -216,7 +215,7 @@ export class BulkActionsRoutes {
             };
 
             for (let modVersion of modVersionFiltered) {
-                let outObj = await modVersion.modVersion.addGameVersionId(gameVersion2.id, session.user.id).catch((err) => {
+                let outObj = await modVersion.modVersion.addGameVersionId(gameVersion2.id, session.user).catch((err) => {
                     Logger.error(`Error adding game version ${gameVersion2.id} to mod version ${modVersion.modVersion.id}: ${err}`);
                     //results.errorIds.push(modVersion.modVersion.id);
                     return null;
