@@ -1,9 +1,8 @@
 import { Router } from 'express';
-import { DatabaseHelper, Status, ModAPIPublicResponse, GameVersion, UserRoles, User, SupportedGames } from '../../shared/Database';
-import { Validator } from '../../shared/Validator';
-import { validateSession } from '../../shared/AuthHelper';
-import { Config } from '../../shared/Config';
-import { Logger } from '../../shared/Logger';
+import { DatabaseHelper, Status, ModAPIPublicResponse, GameVersion, ModVersionAPIPublicResponse } from '../../shared/Database.ts';
+import { Validator } from '../../shared/Validator.ts';
+import { validateSession } from '../../shared/AuthHelper.ts';
+import { Logger } from '../../shared/Logger.ts';
 import { SemVer } from 'semver';
 
 export class GetModRoutes {
@@ -41,20 +40,24 @@ export class GetModRoutes {
                 });
             }
 
-            // only show approved or unverified mods
-            if (reqQuery.data.status !== Status.Verified && reqQuery.data.status !== Status.Unverified) {
-                return res.status(400).send({ message: `Invalid status.` });
-            }
-
             let gameVersion = reqQuery.data.gameVersion ? DatabaseHelper.cache.gameVersions.find((gameVersion) => gameVersion.version === reqQuery.data.gameVersion && gameVersion.gameName === reqQuery.data.gameName) : null;
 
             if (gameVersion === undefined) {
                 return res.status(400).send({ message: `Invalid game version.` });
             }
 
-            let showUnverified = reqQuery.data.status !== `verified`;
-            let statuses = showUnverified ? [Status.Verified, Status.Unverified] : [Status.Verified];
-            let mods: {mod: ModAPIPublicResponse, latest: any}[] = [];
+            let statuses: Status[] = [Status.Verified];
+            switch (reqQuery.data.status) {
+                case `all`:
+                case Status.Unverified:
+                    statuses = [Status.Verified, Status.Unverified, Status.Pending];
+                    break;
+                case Status.Verified:
+                    statuses = [Status.Verified];
+                    break;
+            }
+                
+            let mods: {mod: ModAPIPublicResponse, latest: ModVersionAPIPublicResponse | null}[] = [];
             if (gameVersion === null) {
                 let modDb = DatabaseHelper.cache.mods.filter((mod) => mod.gameName == reqQuery.data.gameName && statuses.includes(mod.status));
 
@@ -68,7 +71,7 @@ export class GetModRoutes {
                     let latest = modVersions[0];
 
                     if (latest) {
-                        let latestVer = await latest.toAPIResonse(latest.supportedGameVersionIds[0], statuses);
+                        let latestVer = await latest.toAPIResponse(latest.supportedGameVersionIds[0], statuses);
                         if (latestVer) {
                             mods.push({ mod: mod.toAPIResponse(), latest: latestVer });
                         }
@@ -80,7 +83,7 @@ export class GetModRoutes {
 
                 for (let retMod of modsFromDB) {
                     let mod = retMod.mod.toAPIResponse();
-                    let latest = await retMod.latest.toAPIResonse(gameVersion.id, statuses);
+                    let latest = await retMod.latest.toAPIResponse(gameVersion.id, statuses);
                     mods.push({ mod, latest });
                 }
 
@@ -133,7 +136,7 @@ export class GetModRoutes {
                 raw = false;
             }
 
-            let mod = DatabaseHelper.cache.mods.find((mod) => mod.id === modId.data);
+            let mod = DatabaseHelper.mapCache.mods.get(modId.data);
             if (!mod) {
                 return res.status(404).send({ message: `Mod not found.` });
             }
@@ -153,9 +156,10 @@ export class GetModRoutes {
                 }
                 // if raw is true, return the raw mod version info instead of attempting to resolve the dependencies & other fields
                 if (raw) {
-                    returnVal.push(version.toRawAPIResonse());
+                    returnVal.push(version.toRawAPIResponse());
                 } else {
-                    let resolvedVersion = await version.toAPIResonse(version.supportedGameVersionIds[0], [Status.Verified, Status.Unverified, Status.Private, Status.Removed]);
+                    // resort to default behavior, which does return no matter what iirc.
+                    let resolvedVersion = await version.toAPIResponse(undefined, [Status.Verified, Status.Unverified, Status.Private, Status.Removed]);
                     if (resolvedVersion) {
                         returnVal.push(resolvedVersion);
                     } else {
@@ -191,28 +195,77 @@ export class GetModRoutes {
             let session = await validateSession(req, res, false, null, false);
             let modVersionId = Validator.zDBID.safeParse(req.params.modVersionIdParam);
             let raw = req.query.raw;
-            if (!modVersionId) {
+            if (!modVersionId.success) {
                 return res.status(400).send({ message: `Invalid mod version id.` });
             }
 
-            let modVersion = DatabaseHelper.cache.modVersions.find((modVersion) => modVersion.id === modVersionId.data);
+            let modVersion = DatabaseHelper.mapCache.modVersions.get(modVersionId.data);
             if (!modVersion) {
                 return res.status(404).send({ message: `Mod version not found.` });
             }
 
-            let mod = DatabaseHelper.cache.mods.find((mod) => mod.id === modVersion.modId);
+            let mod = DatabaseHelper.mapCache.mods.get(modVersion.modId);
             
             if (!await modVersion.isAllowedToView(session.user, mod)) {
                 return res.status(404).send({ message: `Mod version not found.` });
             }
 
             if (raw === `true`) {
-                return res.status(200).send({ mod: mod ? mod.toAPIResponse() : undefined, modVersion: modVersion.toRawAPIResonse() });
+                return res.status(200).send({ mod: mod ? mod.toAPIResponse() : undefined, modVersion: modVersion.toRawAPIResponse() });
             } else {
-                return res.status(200).send({ mod: mod ? mod.toAPIResponse() : undefined, modVersion: await modVersion.toAPIResonse(modVersion.supportedGameVersionIds[0], [Status.Verified, Status.Unverified]) });
+                return res.status(200).send({ mod: mod ? mod.toAPIResponse() : undefined, modVersion: await modVersion.toAPIResponse(modVersion.supportedGameVersionIds[0], [Status.Verified, Status.Unverified]) });
             }
         });
 
+        this.router.get(`/multi/modversions`, async (req, res) => {
+            // #swagger.tags = ['Mods']
+            /* #swagger.security = [{
+                "bearerAuth": [],
+                "cookieAuth": []
+            }] */
+            // #swagger.summary = 'Get multiple mod versions by ID.'
+            // #swagger.description = 'Get multiple mod versions by ID.'
+            // #swagger.responses[200] = { description: 'Returns the mod version.' }
+            // #swagger.responses[400] = { description: 'Invalid mod version id.' }
+            // #swagger.responses[404] = { description: 'Mod version not found.' }
+            // #swagger.parameters['modVersionIds'] = { in: 'query', description: 'The mod version IDs.', type: 'array', required: true }
+            // #swagger.parameters['raw'] = { description: 'Return the raw mod depedendcies without attempting to resolve them.', type: 'boolean' }
+            let session = await validateSession(req, res, false, null, false);
+            let modVersionIds = Validator.zDBIDArray.safeParse(req.query.id);
+            let raw = req.query.raw;
+            if (!modVersionIds.success) {
+                return res.status(400).send({ message: `Invalid mod version id.` });
+            }
+
+            let dedupedIds = Array.from(new Set(modVersionIds.data));
+
+            let retVal: {mod: ModAPIPublicResponse, modVersion: any}[] = [];
+            for (const id of dedupedIds) {
+                let modVersion = DatabaseHelper.mapCache.modVersions.get(id);
+                if (!modVersion) {
+                    return res.status(404).send({ message: `Mod version not found.` });
+                }
+
+                let mod = DatabaseHelper.mapCache.mods.get(modVersion.modId);
+                if (!mod) {
+                    return res.status(404).send({ message: `Mod ID ${modVersion.modId} not found.` });
+                }
+                
+                if (!await modVersion.isAllowedToView(session.user, mod)) {
+                    return res.status(404).send({ message: `Mod version not found.` });
+                }
+
+                if (raw === `true`) {
+                    retVal.push({ mod: mod.toAPIResponse(), modVersion: modVersion.toRawAPIResponse() });
+                } else {
+                    retVal.push({ mod: mod.toAPIResponse(), modVersion: await modVersion.toAPIResponse(modVersion.supportedGameVersionIds[0], [Status.Verified, Status.Unverified, Status.Pending, Status.Removed]) });
+                }
+            }
+
+            return res.status(200).send({ mods: retVal });
+        });
+
+        // #region Hashes
         this.router.get(`/hashlookup`, async (req, res) => {
             // #swagger.tags = ['Mods']
             // #swagger.summary = 'Get a specific mod version that has a file with the specified hash.'
@@ -222,9 +275,10 @@ export class GetModRoutes {
             // #swagger.responses[404] = { description: 'Hash not found.' }
             // #swagger.parameters['hash'] = { description: 'The hash to look up.', type: 'string', required: true }
             // #swagger.parameters['raw'] = { description: 'Return the raw mod depedendcies without attempting to resolve them.', type: 'boolean' }
-
+            // #swagger.parameters['status'] = { description: 'Only show mods with these statuses.', type: 'string' }
             const hash = Validator.zHashStringOrArray.safeParse(req.query.hash).data;
             const raw = Validator.zBool.safeParse(req.query.raw).data;
+            const status = Validator.zStatus.safeParse(req.query.status).data;
 
             if (!hash) {
                 return res.status(400).send({ message: `Missing hash.` });
@@ -235,19 +289,23 @@ export class GetModRoutes {
             let hashArr = Array.isArray(hash) ? hash : [hash];
 
             for (const version of DatabaseHelper.cache.modVersions) {
+                if (status !== undefined && status !== version.status) {
+                    continue;
+                }
+
                 if (hashArr.includes(version.zipHash)) {
                     if (raw) {
-                        retVal.push(Promise.resolve(version.toRawAPIResonse()));
+                        retVal.push(Promise.resolve(version.toRawAPIResponse()));
                     } else {
-                        retVal.push(version.toAPIResonse(version.supportedGameVersionIds[0], [Status.Verified, Status.Unverified]));
+                        retVal.push(version.toAPIResponse(version.supportedGameVersionIds[0], [Status.Verified, Status.Unverified]));
                     }
                 }
                 for (const fileHash of version.contentHashes) {
                     if (hashArr.includes(fileHash.hash)) {
                         if (raw) {
-                            retVal.push(Promise.resolve(version.toRawAPIResonse()));
+                            retVal.push(Promise.resolve(version.toRawAPIResponse()));
                         } else {
-                            retVal.push(version.toAPIResonse(version.supportedGameVersionIds[0], [Status.Verified, Status.Unverified]));
+                            retVal.push(version.toAPIResponse(version.supportedGameVersionIds[0], [Status.Verified, Status.Unverified]));
                         }
                     }
                 }
@@ -262,7 +320,7 @@ export class GetModRoutes {
             }
         });
 
-        this.router.get(`/multihashlookup`, async (req, res) => {
+        this.router.get(`/multi/hashlookup`, async (req, res) => {
             // #swagger.tags = ['Mods']
             // #swagger.summary = 'Get a specific mod version that has a file with the specified hash.'
             // #swagger.description = 'Look up multiple hashes at once, and sort the results by hash. Developed for PinkModManager.'
@@ -270,8 +328,8 @@ export class GetModRoutes {
             // #swagger.responses[400] = { description: 'Missing hash.' }
             // #swagger.responses[404] = { description: 'Hash not found.' }
             // #swagger.parameters['hash'] = { description: 'The hash to look up.', type: 'string', required: true }
-
             const hash = Validator.zHashStringOrArray.safeParse(req.query.hash).data;
+            const status = Validator.zStatus.safeParse(req.query.status).data;
 
             if (!hash) {
                 return res.status(400).send({ message: `Missing hash.` });
@@ -282,21 +340,24 @@ export class GetModRoutes {
             let hashArr = Array.isArray(hash) ? hash : [hash];
 
             for (const version of DatabaseHelper.cache.modVersions) {
+                if (status !== undefined && status !== version.status) {
+                    continue;
+                }
                 if (hashArr.includes(version.zipHash)) {
                     let existing = retVal.get(version.zipHash);
                     if (existing) {
-                        existing.push(version.toRawAPIResonse());
+                        existing.push(version.toRawAPIResponse());
                     } else {
-                        retVal.set(version.zipHash, [version.toRawAPIResonse()]);
+                        retVal.set(version.zipHash, [version.toRawAPIResponse()]);
                     }
                 }
                 for (const fileHash of version.contentHashes) {
                     if (hashArr.includes(fileHash.hash)) {
                         let existing = retVal.get(fileHash.hash);
                         if (existing) {
-                            existing.push(version.toRawAPIResonse());
+                            existing.push(version.toRawAPIResponse());
                         } else {
-                            retVal.set(fileHash.hash, [version.toRawAPIResonse()]);
+                            retVal.set(fileHash.hash, [version.toRawAPIResponse()]);
                         }
                     }
                 }
@@ -309,7 +370,7 @@ export class GetModRoutes {
                     retObj[key] = value;
                 });
 
-                return res.status(200).send({ modVersions: retObj });
+                return res.status(200).send({ hashes: retObj });
             } else {
                 return res.status(404).send({ message: `Hash not found.` });
             }
