@@ -1,12 +1,12 @@
 import { Router } from 'express';
-import { DatabaseHelper, Status, UserRoles } from '../../shared/Database';
-import { validateSession } from '../../shared/AuthHelper';
-import { Logger } from '../../shared/Logger';
-import { Validator } from '../../shared/Validator';
+import { DatabaseHelper, Status } from '../../shared/Database.ts';
+import { validateSession } from '../../shared/AuthHelper.ts';
+import { Logger } from '../../shared/Logger.ts';
+import { Validator } from '../../shared/Validator.ts';
 import { SemVer } from 'semver';
 import path from 'node:path';
-import { Config } from '../../shared/Config';
-import { sendEditLog } from '../../shared/ModWebhooks';
+import { Config } from '../../shared/Config.ts';
+import { Utils } from '../../shared/Utils.ts';
 
 export class UpdateModRoutes {
     private router: Router;
@@ -58,28 +58,8 @@ export class UpdateModRoutes {
                 return res.status(404).send({ message: `Mod not found.` });
             }
 
-            // check permissions
-            let allowedToEdit = false;
-            if (session.user.roles.sitewide.includes(UserRoles.AllPermissions) || session.user.roles.sitewide.includes(UserRoles.Approver) || mod.authorIds.includes(session.user.id)) {
-                allowedToEdit = true;
-            }
-
-            if (reqBody.data.gameName && reqBody.data.gameName !== mod.gameName) {
-                // if changing game, check if user has permissions for the new game
-                if (session.user.roles.perGame[reqBody.data.gameName]?.includes(UserRoles.AllPermissions) || session. user.roles.perGame[reqBody.data.gameName]?.includes(UserRoles.Approver)) {
-                    // and if they have permissions for the current game
-                    if (session.user.roles.perGame[mod.gameName]?.includes(UserRoles.AllPermissions) || session.user.roles.perGame[mod.gameName]?.includes(UserRoles.Approver)) {
-                        allowedToEdit = true;
-                    }
-                }
-            } else {
-                // if not changing game, check if user has permissions for the current game
-                if (session.user.roles.perGame[mod.gameName]?.includes(UserRoles.AllPermissions) || session.user.roles.perGame[mod.gameName]?.includes(UserRoles.Approver)) {
-                    allowedToEdit = true;
-                }
-            }
-
-            if (!allowedToEdit) {
+            let isGameChange = reqBody.data.gameName && reqBody.data.gameName !== mod.gameName;
+            if (mod.isAllowedToEdit(session.user, isGameChange) == false) {
                 return res.status(401).send({ message: `You cannot edit this mod.` });
             }
 
@@ -90,55 +70,31 @@ export class UpdateModRoutes {
                 }
             }
 
-            if (mod.status == Status.Verified) {
-                let existingEdit = await DatabaseHelper.database.EditApprovalQueue.findOne({ where: { objectId: mod.id, objectTableName: `mods`, submitterId: session.user.id, approved: null } });
-
-                if (existingEdit) {
-                    return res.status(400).send({ message: `You already have a pending edit for this mod.` }); // todo: allow updating the edit
-                }
-
-                await DatabaseHelper.database.EditApprovalQueue.create({
-                    submitterId: session.user.id,
-                    objectTableName: `mods`,
-                    objectId: mod.id,
-                    object: {
-                        name: reqBody.data.name || mod.name,
-                        summary: reqBody.data.summary || mod.summary,
-                        description: reqBody.data.description || mod.description,
-                        gameName: reqBody.data.gameName || mod.gameName,
-                        gitUrl: reqBody.data.gitUrl || mod.gitUrl,
-                        authorIds: reqBody.data.authorIds || mod.authorIds,
-                        category: reqBody.data.category || mod.category,
+            mod.edit({
+                name: reqBody.data.name || mod.name,
+                summary: reqBody.data.summary || mod.summary,
+                description: reqBody.data.description || mod.description,
+                gameName: reqBody.data.gameName || mod.gameName,
+                gitUrl: reqBody.data.gitUrl || mod.gitUrl,
+                authorIds: reqBody.data.authorIds || mod.authorIds,
+                category: reqBody.data.category || mod.category,
+            }, session.user).then((mod) => {
+                if (mod.isEditObj) {
+                    if (mod.newEdit) {
+                        res.status(200).send({ message: `Edit ${mod.edit.id} (for ${mod.edit.objectId}) submitted by ${session.user.id} for approval.`, edit: mod.edit });
+                    } else {
+                        res.status(200).send({ message: `Edit ${mod.edit.id} (for ${mod.edit.objectId}) updated by ${session.user.id}.`, edit: mod.edit });
                     }
-                }).then((edit) => {
-                    Logger.log(`Edit ${edit.id} (for ${edit.objectId}) submitted by ${session.user.id} for approval.`);
-                    res.status(200).send({ message: `Edit ${edit.id} (for ${edit.objectId}) submitted by ${session.user.id} for approval.`, edit: edit });
                     DatabaseHelper.refreshCache(`editApprovalQueue`);
-                    sendEditLog(edit, session.user, `New`);
-                }).catch((error) => {
-                    Logger.error(`Error submitting edit: ${error}`);
-                    res.status(500).send({ message: `Error creating edit submitted by ${session.user.id}.` });
-                });
-            } else {
-                await mod.update({
-                    name: reqBody.data.name || mod.name,
-                    summary: reqBody.data.summary || mod.summary,
-                    description: reqBody.data.description || mod.description,
-                    gameName: reqBody.data.gameName || mod.gameName,
-                    gitUrl: reqBody.data.gitUrl || mod.gitUrl,
-                    authorIds: reqBody.data.authorIds || mod.authorIds,
-                    category: reqBody.data.category || mod.category,
-                    lastUpdatedById: session.user.id,
-                }).then((mod) => {
-                    res.status(200).send({ message: `Mod ${mod.id} updated by ${session.user.id}.`, edit: mod });
-                }).catch((error) => {
-                    let message = `Error updating mod.`;
-                    if (Array.isArray(error?.errors) && error?.errors?.length > 0) {
-                        message = error.errors.map((e: any) => e.message).join(`, `);
-                    }
-                    res.status(500).send({ message: `Error updating mod: ${error} ${message} ${error?.name}` });
-                });
-            }
+                    return;
+                } else {
+                    res.status(200).send({ message: `Mod updated.`, mod: mod.mod });
+                    DatabaseHelper.refreshCache(`mods`);
+                }
+            }).catch((error) => {
+                let errorMessage = Utils.parseErrorMessage(error);
+                res.status(500).send({ message: `Error updating mod: ${errorMessage}` });
+            });
         });
 
         this.router.post(`/mods/:modIdParam/icon`, async (req, res) => {
@@ -261,16 +217,7 @@ export class UpdateModRoutes {
                 return res.status(404).send({ message: `Mod not found.` });
             }
 
-            let allowedToEdit = false;
-            if (session.user.roles.sitewide.includes(UserRoles.AllPermissions) || session.user.roles.sitewide.includes(UserRoles.Approver) || mod.authorIds.includes(session.user.id)) {
-                allowedToEdit = true;
-            }
-
-            if (session.user.roles.perGame[mod.gameName]?.includes(UserRoles.AllPermissions) || session.user.roles.perGame[mod.gameName]?.includes(UserRoles.Approver)) {
-                allowedToEdit = true;
-            }
-
-            if (!allowedToEdit) {
+            if (await modVersion.isAllowedToEdit(session.user, mod) == false) {
                 return res.status(401).send({ message: `You cannot edit this mod.` });
             }
 
@@ -286,48 +233,28 @@ export class UpdateModRoutes {
                 }
             }
 
-            if (modVersion.status == Status.Verified) {
-                let existingEdit = await DatabaseHelper.database.EditApprovalQueue.findOne({ where: { objectId: modVersion.id, objectTableName: `modVersions`, submitterId: session.user.id, approved: null } });
-
-                if (existingEdit) {
-                    return res.status(400).send({ message: `You already have a pending edit for this mod version.` }); // todo: allow updating the edit
-                }
-
-                await DatabaseHelper.database.EditApprovalQueue.create({
-                    submitterId: session.user.id,
-                    objectTableName: `modVersions`,
-                    objectId: modVersion.id,
-                    object: {
-                        supportedGameVersionIds: reqBody.data.supportedGameVersionIds || modVersion.supportedGameVersionIds,
-                        modVersion: reqBody.data.modVersion ? new SemVer(reqBody.data.modVersion) : modVersion.modVersion,
-                        dependencies: reqBody.data.dependencies || modVersion.dependencies,
-                        platform: reqBody.data.platform || modVersion.platform,
+            modVersion.edit({
+                supportedGameVersionIds: reqBody.data.supportedGameVersionIds || modVersion.supportedGameVersionIds,
+                modVersion: reqBody.data.modVersion ? new SemVer(reqBody.data.modVersion) : modVersion.modVersion,
+                dependencies: reqBody.data.dependencies || modVersion.dependencies,
+                platform: reqBody.data.platform || modVersion.platform,
+            }, session.user).then((modVersion) => {
+                if (modVersion.isEditObj) {
+                    if (modVersion.newEdit) {
+                        res.status(200).send({ message: `Edit ${modVersion.edit.id} (for ${modVersion.edit.objectId}) submitted by ${session.user.id} for approval.`, edit: modVersion.edit });
+                    } else {
+                        res.status(200).send({ message: `Edit ${modVersion.edit.id} (for ${modVersion.edit.objectId}) updated by ${session.user.id}.`, edit: modVersion.edit });
                     }
-                }).then((edit) => {
-                    res.status(200).send({ message: `Edit ${edit.id} (for ${edit.objectId}) submitted by ${session.user.id} for approval.`, edit: edit });
                     DatabaseHelper.refreshCache(`editApprovalQueue`);
-                    sendEditLog(edit, session.user, `New`);
-                }).catch((error) => {
-                    Logger.error(`Error submitting edit: ${error}`);
-                    res.status(500).send({ message: `Error submitting edit.` });
-                });
-            } else {
-                await modVersion.update({
-                    supportedGameVersionIds: reqBody.data.supportedGameVersionIds || modVersion.supportedGameVersionIds,
-                    modVersion: reqBody.data.modVersion ? new SemVer(reqBody.data.modVersion) : modVersion.modVersion,
-                    dependencies: reqBody.data.dependencies || modVersion.dependencies,
-                    platform: reqBody.data.platform || modVersion.platform,
-                }).then((modVersion) => {
-                    DatabaseHelper.refreshCache(`modVersions`);
+                    return;
+                } else {
                     res.status(200).send({ message: `Mod version updated.`, modVersion });
-                }).catch((error) => {
-                    let message = `Error updating version.`;
-                    if (Array.isArray(error?.errors) && error?.errors?.length > 0) {
-                        message = error.errors.map((e: any) => e.message).join(`, `);
-                    }
-                    res.status(500).send({ message: `Error updating version: ${error} ${message} ${error?.name}` });
-                });
-            }
+                    DatabaseHelper.refreshCache(`modVersions`);
+                }
+            }).catch((error) => {
+                let errorMessage = Utils.parseErrorMessage(error);
+                res.status(500).send({ message: `Error updating mod version: ${errorMessage}` });
+            });
         });
         // #endregion Update Mod Version
         // #region Submit to Approval
@@ -356,7 +283,7 @@ export class UpdateModRoutes {
                 return res.status(400).send({ message: `Mod is already submitted.` });
             }
 
-            mod.setStatus(Status.Unverified, session.user).then((mod) => {
+            mod.setStatus(Status.Pending, session.user, `Mod submitted for verification by ${session.user.username}`).then((mod) => {
                 res.status(200).send({ message: `Mod submitted.`, mod });
                 DatabaseHelper.refreshCache(`mods`);
             }).catch((error) => {
@@ -398,15 +325,16 @@ export class UpdateModRoutes {
                 return res.status(400).send({ message: `Mod version is already submitted.` });
             }
 
-            modVersion.setStatus(Status.Unverified, session.user).then((modVersion) => {
+            modVersion.setStatus(Status.Pending, session.user, `Version submitted for approval by ${session.user.username}`).then((modVersion) => {
                 res.status(200).send({ message: `Mod version submitted.`, modVersion });
                 DatabaseHelper.refreshCache(`modVersions`);
             }).catch((error) => {
                 res.status(500).send({ message: `Error submitting mod version: ${error}` });
             });
-            // #endregion Submit
+            
         });
-
+        // #endregion Submit
+        // #region Edits
         this.router.get(`/edits`, async (req, res) => {
             // #swagger.tags = ['Mods']
             /* #swagger.security = [{
@@ -427,7 +355,7 @@ export class UpdateModRoutes {
                 if (edit.isMod()) {
                     return edit.submitterId == session.user.id || usersMods.some((mod) => edit.objectId == mod.id);
                 } else {
-                    let modVersion = DatabaseHelper.cache.modVersions.find((mod) => mod.id == edit.objectId);
+                    let modVersion = DatabaseHelper.mapCache.modVersions.get(edit.objectId);
                     if (!modVersion) {
                         return false;
                     }
@@ -461,7 +389,7 @@ export class UpdateModRoutes {
                 return res.status(404).send({ message: `Edit not found.` });
             }
 
-            let parentObj = edit.isMod() ? DatabaseHelper.cache.mods.find((mod) => mod.id == edit.objectId) : DatabaseHelper.cache.modVersions.find((modVersion) => modVersion.id == edit.objectId);
+            let parentObj = edit.isMod() ? DatabaseHelper.mapCache.mods.get(edit.objectId) : DatabaseHelper.mapCache.modVersions.get(edit.objectId);
             if (!parentObj) {
                 return res.status(404).send({ message: `Parent object not found.` });
             }
@@ -473,6 +401,49 @@ export class UpdateModRoutes {
             }
 
             return res.status(200).send({ message: `Edit found.`, edit });
+        });
+
+        this.router.delete(`/edits/:editIdParam`, async (req, res) => {
+            // #swagger.tags = ['Mods']
+            /* #swagger.security = [{
+                "bearerAuth": [],
+                "cookieAuth": []
+            }] */
+            // #swagger.description = `Delete an edit.`
+            // #swagger.parameters['editIdParam'] = { description: 'Edit ID', type: 'integer' }
+            let editId = Validator.zDBID.safeParse(req.params.editIdParam);
+            if (!editId.success) {
+                return res.status(400).send({ message: `Invalid editId.` });
+            }
+
+            let session = await validateSession(req, res, true);
+            if (!session.user) {
+                return;
+            }
+
+            let edit = DatabaseHelper.cache.editApprovalQueue.find((edit) => edit.id == editId.data);
+            if (!edit) {
+                return res.status(404).send({ message: `Edit not found.` });
+            }
+
+            let parentObj = edit.isMod() ? DatabaseHelper.mapCache.mods.get(edit.objectId) : DatabaseHelper.mapCache.modVersions.get(edit.objectId);
+            if (!parentObj) {
+                return res.status(404).send({ message: `Parent object not found.` });
+            }
+
+            let isAllowedToDelete = parentObj?.isAllowedToEdit(session.user);
+
+            if (!isAllowedToDelete) {
+                return res.status(401).send({ message: `You cannot delete this edit.` });
+            }
+
+            await edit.deny(session.user).then(() => {
+                res.status(200).send({ message: `Edit deleted.` });
+                DatabaseHelper.refreshCache(`editApprovalQueue`);
+            }).catch((error) => {
+                let errorMessage = Utils.parseErrorMessage(error);
+                res.status(500).send({ message: `Error deleting edit: ${errorMessage}` });
+            });
         });
     }
 }
