@@ -1,3 +1,5 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+/* eslint-disable no-console */
 import express, { Request } from 'express';
 import session, { SessionOptions } from 'express-session';
 import MemoryStore from 'memorystore';
@@ -11,129 +13,138 @@ import fs from 'node:fs';
 import { ActivityType } from 'discord.js';
 import passport from 'passport';
 import { Strategy as BearerStrategy } from 'passport-http-bearer';
-// eslint-disable-next-line @typescript-eslint/no-var-requires
+import { Octokit } from '@octokit/rest';
 
-import { DatabaseHelper, DatabaseManager } from './shared/Database';
-import { Logger } from './shared/Logger';
-import { Config } from './shared/Config';
-import { Luma } from './discord/classes/Luma';
+import { DatabaseHelper, DatabaseManager } from './shared/Database.ts';
+import { Logger } from './shared/Logger.ts';
+import { Config } from './shared/Config.ts';
+import { Luma } from './discord/classes/Luma.ts';
 
-import { CreateModRoutes } from './api/routes/createMod';
-import { GetModRoutes } from './api/routes/getMod';
-import { UpdateModRoutes } from './api/routes/updateMod';
-import { AuthRoutes } from './api/routes/auth';
-import { VersionsRoutes } from './api/routes/versions';
-import { ImportRoutes } from './api/routes/import';
-import { AdminRoutes } from './api/routes/admin';
-import { ApprovalRoutes } from './api/routes/approval';
-import { BeatModsRoutes } from './api/routes/beatmods';
-import { CDNRoutes } from './api/routes/cdn';
-import { MOTDRoutes } from './api/routes/motd';
-import { UserRoutes } from './api/routes/users';
-import { StatusRoutes } from './api/routes/status';
-import { BulkActionsRoutes } from './api/routes/bulkActions';
+import { CreateModRoutes } from './api/routes/createMod.ts';
+import { GetModRoutes } from './api/routes/getMod.ts';
+import { UpdateModRoutes } from './api/routes/updateMod.ts';
+import { AuthRoutes } from './api/routes/auth.ts';
+import { VersionsRoutes } from './api/routes/versions.ts';
+import { ImportRoutes } from './api/routes/import.ts';
+import { AdminRoutes } from './api/routes/admin.ts';
+import { ApprovalRoutes } from './api/routes/approval.ts';
+import { BeatModsRoutes } from './api/routes/beatmods.ts';
+import { CDNRoutes } from './api/routes/cdn.ts';
+import { MOTDRoutes } from './api/routes/motd.ts';
+import { UserRoutes } from './api/routes/users.ts';
+import { StatusRoutes } from './api/routes/apistatus.ts';
+import { BulkActionsRoutes } from './api/routes/bulkActions.ts';
 
-import swaggerDocument from './api/swagger.json';
-// eslint-disable-next-line no-console
-console.log(`Starting setup...`);
-new Config();
-new Logger();
-const app = express();
-const memstore = MemoryStore(session);
-const port = Config.server.port;
-let database = new DatabaseManager();
-
-// handle parsing request bodies
-app.use(express.json({ limit: 100000 }));
-app.use(express.urlencoded({limit : 10000, parameterLimit: 10, extended: false }));
-app.use(cors({
-    origin: Config.server.corsOrigins,
-    credentials: Config.server.iHateSecurity ? true : false,
-}));
-app.use(fileUpload({
-    limits: {
-        fileSize: Math.floor(Config.server.fileUploadLimitMB * Config.server.fileUploadMultiplierMB * 1024 * 1024), // here you go kaitlyn
-        files: 1
-    },
-    abortOnLimit: true,
-    limitHandler: (req, res, next) => {
-        return res.status(413).send({ message: `File size limit has been reached.` });
-    },
-}));
-
-const sessionConfigData: SessionOptions = {
-    secret: Config.server.sessionSecret,
-    name: `bbm_session`,
-    resave: false,
-    saveUninitialized: false,
-    unset: `destroy`,
-    rolling: true,
-    cookie: {
-        maxAge: 1000 * 60 * 60 * 24 * 7, // 1 week
-        secure: `auto`,
-        httpOnly: true,
-        sameSite: Config.server.iHateSecurity ? `none` : `strict`,
+// eslint-disable-next-line quotes
+import fullApi from './api/swagger_full.json' with { type: "json" };
+// eslint-disable-next-line quotes
+import publicApi from './api/swagger_public.json' with { type: "json" };
+import { Server } from 'node:http';
+function init() {
+    console.log(`Starting setup...`);
+    if (process.env.NODE_ENV === `test`) {
+        console.log(`Running in test mode.`);
+    } else {
+        new Config();
     }
-};
+    new Logger();
+    const app = express();
+    const memstore = MemoryStore(session);
+    const port = Config.server.port;
+    let database = new DatabaseManager();
+    let server: Server | undefined = undefined;
+    let bot: Luma | undefined = undefined;
 
-if (Config.server.storeSessions) {
-    const sqlite3sessions = connectSqlite3(session);
-    let dbpath = Config.storage.sessions.split(`/`);
-    let name = dbpath.pop();
-    if (name === undefined) {
-        throw new Error(`Invalid session storage path.`);
+    // handle parsing request bodies
+    app.use(express.json({ limit: 100000 }));
+    app.use(express.urlencoded({limit : 10000, parameterLimit: 10, extended: false }));
+    app.use(cors({
+        origin: Config.server.corsOrigins,
+        credentials: Config.server.iHateSecurity ? true : false,
+    }));
+    app.use(fileUpload({
+        limits: {
+            fileSize: Math.floor(Config.server.fileUploadLimitMB * Config.server.fileUploadMultiplierMB * 1024 * 1024), // here you go kaitlyn
+            files: 1
+        },
+        abortOnLimit: true,
+        limitHandler: (req, res, next) => {
+            return res.status(413).send({ message: `File size limit has been reached.` });
+        },
+    }));
+
+    const sessionConfigData: SessionOptions = {
+        secret: Config.server.sessionSecret,
+        name: `bbm_session`,
+        resave: false,
+        saveUninitialized: false,
+        unset: `destroy`,
+        rolling: true,
+        cookie: {
+            maxAge: 1000 * 60 * 60 * 24 * 7, // 1 week
+            secure: `auto`,
+            httpOnly: true,
+            sameSite: Config.server.iHateSecurity ? `none` : `strict`,
+        }
+    };
+
+    if (Config.server.storeSessions) {
+        const sqlite3sessions = connectSqlite3(session);
+        let dbpath = Config.storage.sessions.split(`/`);
+        let name = dbpath.pop();
+        if (name === undefined) {
+            throw new Error(`Invalid session storage path.`);
+        }
+        name = name.split(`.`)[0];
+        sessionConfigData.store = new (sqlite3sessions as any)({
+            db: name,
+            dir: path.resolve(dbpath.join(`/`)),
+            table: `sessions`
+        });
+    } else {
+        sessionConfigData.store = new memstore({
+            checkPeriod: 86400000,
+        });
     }
-    name = name.split(`.`)[0];
-    sessionConfigData.store = new (sqlite3sessions as any)({
-        db: name,
-        dir: path.resolve(dbpath.join(`/`)),
-        table: `sessions`
+
+    app.set(`trust proxy`, Config.server.trustProxy);
+
+    let apiRouter = express.Router({
+        caseSensitive: false,
+        mergeParams: false,
+        strict: false,
     });
-} else {
-    sessionConfigData.store = new memstore({
-        checkPeriod: 86400000,
+
+    let cdnRouter = express.Router({
+        caseSensitive: false,
+        mergeParams: false,
+        strict: false,
     });
-}
 
-app.set(`trust proxy`, Config.server.trustProxy);
+    apiRouter.use(rateLimit({
+        windowMs: 60 * 1000,
+        max: 100,
+        statusCode: 429,
+        message: {message: `Rate limit exceeded.`},
+        skipSuccessfulRequests: false,
+        validate: {trustProxy: false},
+    }));
 
-let apiRouter = express.Router({
-    caseSensitive: false,
-    mergeParams: false,
-    strict: false,
-});
+    const cdnRateLimiter = rateLimit({
+        windowMs: 60 * 1000,
+        max: 100,
+        statusCode: 429,
+        message: `Rate limit exceeded.`,
+        skipSuccessfulRequests: false,
+        validate: {trustProxy: false},
+    });
 
-let cdnRouter = express.Router({
-    caseSensitive: false,
-    mergeParams: false,
-    strict: false,
-});
+    cdnRouter.use(cdnRateLimiter);
 
-apiRouter.use(rateLimit({
-    windowMs: 60 * 1000,
-    max: 200,
-    statusCode: 429,
-    message: {message: `Rate limit exceeded.`},
-    skipSuccessfulRequests: false,
-    validate: {trustProxy: false},
-}));
-
-const cdnRateLimiter = rateLimit({
-    windowMs: 60 * 1000,
-    max: 200,
-    statusCode: 429,
-    message: `Rate limit exceeded.`,
-    skipSuccessfulRequests: false,
-    validate: {trustProxy: false},
-});
-
-cdnRouter.use(cdnRateLimiter);
-
-app.use(session(sessionConfigData));
-import(`@octokit/rest`).then((Octokit) => {
+    app.use(session(sessionConfigData));
     passport.use(`bearer`, new BearerStrategy(
         function(token, done) {
-            const octokit = new Octokit.Octokit({ auth: token });
+            const octokit = new Octokit({ auth: token });
             if (invalidAttempts.filter((t) => token === t).length > 2) {
                 return done(null, false);
             }
@@ -164,172 +175,205 @@ import(`@octokit/rest`).then((Octokit) => {
             });
         }
     ));
-});
 
-let invalidAttempts: string[] = [];
-apiRouter.use(async (req, res, next) => {
-    if (req.session.userId || Config.flags.enableGithubPAT == false) {
-        req.bbmAuth = {
-            userId: req.session.userId,
-            isApiAuth: false,
-        };
-        next();
-    } else {
-        passport.authenticate(`bearer`, { session: false }, (err:any, user:any) => {
-            if (err) {
-                return res.status(401).send({ message: `Unauthorized` });
-            }
-            if (user && user.id) {
-                req.bbmAuth = {
-                    userId: user.id,
-                    isApiAuth: true,
-                };
-                //req.session.userId = user.id;
-                //req.session.goodMorning47YourTargetIsThisSession = true;
-            }
+    let invalidAttempts: string[] = [];
+    apiRouter.use(async (req, res, next) => {
+        if (req.session.userId || Config.flags.enableGithubPAT == false) {
+            req.bbmAuth = {
+                userId: req.session.userId,
+                isApiAuth: false,
+            };
             next();
-        })(req, res, next);
-    }
-});
-
-app.use((req, res, next) => {
-    if (Config.devmode) {
-        if (!req.url.includes(`hashlookup`)) {
-            Logger.winston.log(`http`, `${req.method} ${req.url}`);
+        } else {
+            passport.authenticate(`bearer`, { session: false }, (err:any, user:any) => {
+                if (err) {
+                    return res.status(401).send({ message: `Unauthorized` });
+                }
+                if (user && user.id) {
+                    req.bbmAuth = {
+                        userId: user.id,
+                        isApiAuth: true,
+                    };
+                    //req.session.userId = user.id;
+                    //req.session.goodMorning47YourTargetIsThisSession = true;
+                }
+                next();
+            })(req, res, next);
         }
+    });
+
+    app.use((req, res, next) => {
+        if (Config.devmode) {
+            if (!req.url.includes(`hashlookup`)) {
+                Logger.winston.log(`http`, `${req.method} ${req.url}`);
+            }
+        }
+        next();
+    });
+
+    //app.use(`/api`, Validator.runValidator);
+    if (Config.flags.enableBeatModsCompatibility) {
+        new BeatModsRoutes(app, apiRouter);
     }
-    next();
-});
+    new CreateModRoutes(apiRouter);
+    new GetModRoutes(apiRouter);
+    new UpdateModRoutes(apiRouter);
+    new ApprovalRoutes(apiRouter);
+    new AuthRoutes(apiRouter);
+    new ImportRoutes(apiRouter);
+    new AdminRoutes(apiRouter);
+    new VersionsRoutes(apiRouter);
+    new MOTDRoutes(apiRouter);
+    new UserRoutes(apiRouter);
+    new StatusRoutes(apiRouter);
+    new BulkActionsRoutes(apiRouter);
 
-//app.use(`/api`, Validator.runValidator);
-if (Config.flags.enableBeatModsCompatibility) {
-    new BeatModsRoutes(app, apiRouter);
-}
-new CreateModRoutes(apiRouter);
-new GetModRoutes(apiRouter);
-new UpdateModRoutes(apiRouter);
-new ApprovalRoutes(apiRouter);
-new AuthRoutes(apiRouter);
-new ImportRoutes(apiRouter);
-new AdminRoutes(apiRouter);
-new VersionsRoutes(apiRouter);
-new MOTDRoutes(apiRouter);
-new UserRoutes(apiRouter);
-new StatusRoutes(apiRouter);
-new BulkActionsRoutes(apiRouter);
-
-if (Config.flags.enableSwagger) {
-    swaggerDocument.servers = [{url: `${Config.server.url}${Config.server.apiRoute}`}];
-    if (!Config.flags.enableGithubPAT) {
+    if (Config.flags.enableSwagger) {
+        publicApi.servers = [{url: `${Config.server.url}${Config.server.apiRoute}`}];
+        fullApi.servers = [{url: `${Config.server.url}${Config.server.apiRoute}`}];
+        if (!Config.flags.enableGithubPAT) {
         // @ts-expect-error it complains about it not being undefineable. this just in! i dont care.
-        swaggerDocument.components.securitySchemes.bearerAuth = undefined;
-    }
-    apiRouter.use(`/docs`, swaggerUi.serve, swaggerUi.setup(swaggerDocument, {
-        swaggerOptions: {
-            docExpansion: `list`,
-            defaultModelExpandDepth: 2,
-            defaultModelsExpandDepth: 2,
+            publicApi.components.securitySchemes.bearerAuth = undefined;
+            // @ts-expect-error it complains about it not being undefineable.
+            fullApi.components.securitySchemes.bearerAuth = undefined;
         }
-    }));
-}
-
-if (Config.flags.enableFavicon) {
-    app.get(`/favicon.ico`, cdnRateLimiter, (req, res) => {
-        // #swagger.ignore = true;
-        res.sendFile(path.resolve(`./assets/favicon.png`), {
-            maxAge: 1000 * 60 * 60 * 24 * 1,
-            //immutable: true,
-            lastModified: true,
+        apiRouter.get(`/swagger/full.json`, (req, res) => {
+            res.setHeader(`Cache-Control`, `public, max-age=3600`);
+            res.json(fullApi);
         });
-    });
-}
-        
-if (Config.flags.enableBanner) {
-    // #swagger.ignore = true;
-    app.get(`/banner.png`, cdnRateLimiter, (req, res) => {
-        res.sendFile(path.resolve(`./assets/banner.png`), {
-            maxAge: 1000 * 60 * 60 * 24 * 1,
-            //immutable: true,
-            lastModified: true,
+        apiRouter.get(`/swagger/public.json`, (req, res) => {
+            res.setHeader(`Cache-Control`, `public, max-age=3600`);
+            res.json(publicApi);
         });
-    });
-}
 
-if (Config.devmode && fs.existsSync(path.resolve(`./storage/frontend`))) {
-    app.use(`/`, cdnRateLimiter, express.static(path.resolve(`./storage/frontend`), {
-        dotfiles: `ignore`,
-        immutable: false,
-        index: true,
-        maxAge: 1000 * 60 * 60 * 1,
-        fallthrough: true,
-    }));
-}
-
-new CDNRoutes(cdnRouter);
-
-
-app.use(Config.server.apiRoute, apiRouter);
-app.use(Config.server.cdnRoute, cdnRouter);
-
-app.disable(`x-powered-by`);
-// catch all unknown routes and return a 404
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-apiRouter.use((req, res, next) => {
-    return res.status(404).send({message: `Unknown route.`});
-});
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-cdnRouter.use((req, res, next) => {
-    return res.status(404).send({message: `Unknown route.`});
-});
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-apiRouter.use((err:any, req:any, res:any, next:any) => {
-    Logger.error(err.stack);
-    return res.status(500).send({message: `Server error`});
-});
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-cdnRouter.use((err:any, req:any, res:any, next:any) => {
-    Logger.error(err.stack);
-    return res.status(500).send({message: `Server error`});
-});
-
-
-process.on(`exit`, (code) => {
-    Logger.log(`Process exiting with code ${code}`);
-});
-
-process.on(`SIGTERM`, () => {
-    Logger.log(`Received SIGTERM, exiting.`);
-    DatabaseHelper.database.sequelize.close();
-    process.exit(0);
-});
-
-process.on(`SIGINT`, () => {
-    Logger.log(`Received SIGINT, exiting.`);
-    DatabaseHelper.database.sequelize.close();
-    process.exit(0);
-});
-
-process.on(`SIGQUIT`, () => {
-    Logger.log(`Received SIGQUIT, exiting.`);
-    DatabaseHelper.database.sequelize.close();
-    process.exit(0);
-});
-
-process.on(`unhandledRejection`, (reason: Error | any, promise: Promise<any>) => {
-    if (reason instanceof Error) {
-        Logger.error(`Unhandled promise rejection:${reason.name}\n${reason.message}\n${reason.stack}`);
-    } else {
-        Logger.error(`Unhandled promise rejection:${reason}\n`);
+        apiRouter.use(`/docs`, swaggerUi.serve, swaggerUi.setup(undefined, {
+            explorer: true,
+            swaggerOptions: {
+                docExpansion: `list`,
+                defaultModelExpandDepth: 2,
+                defaultModelsExpandDepth: 2,
+                urls: [
+                    {
+                        url: `${Config.server.url}${Config.server.apiRoute}/swagger/full.json`,
+                        name: `Full API`
+                    },
+                    {
+                        url: `${Config.server.url}${Config.server.apiRoute}/swagger/public.json`,
+                        name: `Public API`,
+                    }
+                ]
+            }
+        }));
     }
-    process.exit(1);
-});
 
-Logger.debug(`Setup complete.`);
+    if (Config.flags.enableFavicon) {
+        app.get(`/favicon.ico`, cdnRateLimiter, (req, res) => {
+        // #swagger.ignore = true;
+            res.sendFile(path.resolve(`./assets/favicon.png`), {
+                maxAge: 1000 * 60 * 60 * 24 * 1,
+                //immutable: true,
+                lastModified: true,
+            });
+        });
+    }
+        
+    if (Config.flags.enableBanner) {
+    // #swagger.ignore = true;
+        app.get(`/banner.png`, cdnRateLimiter, (req, res) => {
+            res.sendFile(path.resolve(`./assets/banner.png`), {
+                maxAge: 1000 * 60 * 60 * 24 * 1,
+                //immutable: true,
+                lastModified: true,
+            });
+        });
+    }
 
-async function startServer() {
+    if (Config.devmode && fs.existsSync(path.resolve(`./storage/frontend`))) {
+        app.use(`/`, cdnRateLimiter, express.static(path.resolve(`./storage/frontend`), {
+            dotfiles: `ignore`,
+            immutable: false,
+            index: true,
+            maxAge: 1000 * 60 * 60 * 1,
+            fallthrough: true,
+        }));
+    }
+
+    new CDNRoutes(cdnRouter);
+
+    app.use(Config.server.apiRoute, apiRouter);
+    app.use(Config.server.cdnRoute, cdnRouter);
+
+    app.disable(`x-powered-by`);
+    // catch all unknown routes and return a 404
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    apiRouter.use((req, res, next) => {
+        return res.status(404).send({message: `Unknown route.`});
+    });
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    cdnRouter.use((req, res, next) => {
+        return res.status(404).send({message: `Unknown route.`});
+    });
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    apiRouter.use((err:any, req:any, res:any, next:any) => {
+        Logger.error(err.stack);
+        return res.status(500).send({message: `Server error`});
+    });
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    cdnRouter.use((err:any, req:any, res:any, next:any) => {
+        Logger.error(err.stack);
+        return res.status(500).send({message: `Server error`});
+    });
+
+
+    process.on(`exit`, (code) => {
+        Logger.log(`Process exiting with code ${code}`);
+    });
+
+    process.on(`SIGTERM`, () => {
+        Logger.log(`Received SIGTERM, exiting.`);
+        DatabaseHelper.database.sequelize.close();
+        process.exit(0);
+    });
+
+    process.on(`SIGINT`, () => {
+        Logger.log(`Received SIGINT, exiting.`);
+        DatabaseHelper.database.sequelize.close();
+        process.exit(0);
+    });
+
+    process.on(`SIGQUIT`, () => {
+        Logger.log(`Received SIGQUIT, exiting.`);
+        DatabaseHelper.database.sequelize.close();
+        process.exit(0);
+    });
+
+    
+    process.on(`unhandledRejection`, (reason: Error | any, promise: Promise<any>) => {
+        if (reason instanceof Error) {
+            Logger.error(`Unhandled promise rejection:${reason.name}\n${reason.message}\n${reason.stack}`);
+        } else {
+            Logger.error(`Unhandled promise rejection:${reason}\n`);
+        }
+        if (process.env.NODE_ENV == `test`) {
+            throw reason;
+        } else {
+            process.exit(1);
+        }
+    });
+
+
+    Logger.debug(`Setup complete.`);
+    return {app, database, port};
+}
+
+export async function startServer() {
+    const {app, database, port} = init();
+    if (process.env.NODE_ENV === `test`) {
+        Logger.debug(`Running in test mode.`);
+    }
     await database.init();
     Logger.debug(`Starting server.`);
-    app.listen(port, () => {
+    const server = app.listen(port, () => {
         Logger.log(`Server listening on port ${port} - Expected to be available at ${Config.server.url}`, ``, true);
         Config.devmode ? Logger.warn(`Development mode is enabled!`) : null;
         Config.authBypass ? Logger.warn(`Authentication bypass is enabled!`) : null;
@@ -337,12 +381,29 @@ async function startServer() {
         Logger.log(`Server started.`);
     });
     
+    let luma = undefined;
     if (Config.bot.enabled) {
-        const luma = new Luma({
+        luma = new Luma({
             intents: [],
             presence: {activities: [{name: `with your mods`, type: ActivityType.Playing}], status: `online`}});
         luma.login(Config.bot.token);
     }
 
+    let stopServer = async (doExit = true, code = 0) => {
+        let promises = [];
+        promises.push(database.sequelize.close());
+        server?.closeAllConnections();
+        server?.close();
+        promises.push(luma?.destroy());
+
+        await Promise.all(promises).then(() => {
+            doExit ? process.exit(code) : undefined;
+        });
+    };
+
+    return {app, server, database, stopServer};
 }
-startServer();
+
+if (process.env.NODE_ENV !== `test`) {
+    startServer();
+}
