@@ -1,4 +1,5 @@
-import { Config } from "../shared/Config";
+/* eslint-disable @typescript-eslint/no-unused-vars */ // disabled for the whole file because the functions almost all have unused parameters
+import { Config } from "../shared/Config.ts";
 import * as Winston from "winston";
 import { DiscordTransport } from "winston-transport-discord";
 
@@ -14,6 +15,7 @@ export class Logger {
         if (Config.webhooks.enableWebhooks && Config.webhooks.loggingUrl.length > 8) {
             transports.push(new DiscordTransport({
                 level: `info`,
+                silent: process.env.NODE_ENV == `test`,
                 discord: {
                     webhook: {
                         url: Config.webhooks.loggingUrl,
@@ -22,9 +24,17 @@ export class Logger {
             }));
         }
 
+        let consoleLevel = `consoleInfo`;
+        if (process.env.NODE_ENV == `test`) {
+            consoleLevel = `warn`;
+        } else if (Config.devmode) {
+            consoleLevel = `http`;
+        }
+
         transports.push(new Winston.transports.Console({
             forceConsole: true,
-            level: Config.devmode ? `http` : `consoleInfo`,
+            level: consoleLevel,
+            //silent: process.env.NODE_ENV == `test`,
             consoleWarnLevels: [`consoleWarn`, `warn`, `error`, `debugWarn`],
             format: Winston.format.combine(
                 Winston.format.timestamp({ format: `MM/DD/YY HH:mm:ss` }),
@@ -38,6 +48,7 @@ export class Logger {
             //filename: `storage/logs/${new Date(Date.now()).toLocaleDateString(`en-US`, { year: `numeric`, month: `numeric`, day: `numeric`}).replaceAll(`/`, `-`)}.log`,
             zippedArchive: true,
             maxsize: 20 * 1024 * 1024,
+            silent: process.env.NODE_ENV == `test`,
             maxFiles: Config.flags.enableUnlimitedLogs ? undefined : 14,
             level: Config.devmode ? `debug` : `info`,
             format: Winston.format.combine(
